+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- feat(core): backport contextvars
 - fix(sanic): fix patching for sanic async http server (#1659)
 - fix(flask): make template patching idempotent